FEAT: Convert AddressSpace to use uintptr_t for 64-bit address support
- Change AddressSpace interface from uint32_t to uintptr_t for guestAddress parameters
- Update DirectAddressSpace implementation to handle full 64-bit addresses
- Remove address truncation that was preventing x86-32 on x86-64 host
- Add sanity checks for invalid addresses
- First instruction successfully reads from 64-bit EIP in direct memory mode
- TODO: Fix infinite loop/crash during second instruction decode

Amp-Thread-ID: https://ampcode.com/threads/T-019c2a30-9a2e-756b-bb18-1abb0841ae9a

+++ AddressSpace.h
@@ -6,55 +6,57 @@
 
 #include <SupportDefs.h>
 #include <stdint.h>
+#include <cstddef>
 
 // Interfaz para la gestión de la memoria del invitado.
 // Abstrae si la memoria es un bloque contiguo, paginada, etc.
+// Uses uintptr_t to support both 32-bit and 64-bit guest addresses on 64-bit hosts
 class AddressSpace {
 public:
 	virtual ~AddressSpace() = default;
 
-	virtual status_t Read(uint32_t guestAddress, void* buffer, size_t size) = 0;
-	virtual status_t ReadString(uint32_t guestAddress, char* buffer, size_t bufferSize) = 0;
-	virtual status_t Write(uint32_t guestAddress, const void* buffer, size_t size) = 0;
+	virtual status_t Read(uintptr_t guestAddress, void* buffer, size_t size) = 0;
+	virtual status_t ReadString(uintptr_t guestAddress, char* buffer, size_t bufferSize) = 0;
+	virtual status_t Write(uintptr_t guestAddress, const void* buffer, size_t size) = 0;
 
 	// Virtual address mapping (optional, default implementations do nothing)
-	virtual status_t RegisterMapping(uint32_t guest_vaddr, uint32_t guest_offset, size_t size)
+	virtual status_t RegisterMapping(uintptr_t guest_vaddr, uintptr_t guest_offset, size_t size)
 	{
 		return B_OK;  // Default: ignore mapping
 	}
 	
-	virtual uint32_t TranslateAddress(uint32_t guest_vaddr) const
+	virtual uintptr_t TranslateAddress(uintptr_t guest_vaddr) const
 	{
 		return guest_vaddr;  // Default: no translation
 	}
 	
 	// TLS area setup (optional)
-	virtual status_t MapTLSArea(uint32_t guest_vaddr, size_t size)
+	virtual status_t MapTLSArea(uintptr_t guest_vaddr, size_t size)
 	{
 		return B_OK;  // Default: ignore
 	}
 	
 	// Direct memory operations (optional, can be overridden)
-	virtual status_t ReadMemory(uint32_t guestAddress, void* data, size_t size)
+	virtual status_t ReadMemory(uintptr_t guestAddress, void* data, size_t size)
 	{
 		return Read(guestAddress, data, size);
 	}
 	
-	virtual status_t WriteMemory(uint32_t guestAddress, const void* data, size_t size)
+	virtual status_t WriteMemory(uintptr_t guestAddress, const void* data, size_t size)
 	{
 		return Write(guestAddress, data, size);
 	}
 
 	// Helper para leer un solo valor
 	template<typename T>
-	status_t Read(uint32_t guestAddress, T& value)
+	status_t Read(uintptr_t guestAddress, T& value)
 	{
 		return Read(guestAddress, &value, sizeof(T));
 	}
 
 	// Helper para escribir un solo valor
 	template<typename T>
-	status_t Write(uint32_t guestAddress, const T& value)
+	status_t Write(uintptr_t guestAddress, const T& value)
 	{
 		return Write(guestAddress, &value, sizeof(T));
 	}
